third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/notifications/serviceworkerregistration-document-data-throw-expected.txt

This is a testharness.js-based test.
PASS showNotification() must reject if data is invalid value. 
Harness: the test ran to completion.
